// src/enums/TargetAudience.ts
export enum TargetAudience {
  ALL = "ALL", // Herkese
  NEW_STUDENTS = "NEW_STUDENTS", // Yeni öğrenciler
  EXISTING_STUDENTS = "EXISTING_STUDENTS", // Mevcut öğrenciler
  SIBLINGS = "SIBLINGS", // Kardeşler
  EARLY_REGISTRANTS = "EARLY_REGISTRANTS", // Erken kayıt olanlar
  LOCAL_RESIDENTS = "LOCAL_RESIDENTS", // Yerel sakinler
  REFERRALS = "REFERRALS", // Tavsiye edilenler
  VIP_MEMBERS = "VIP_MEMBERS", // VIP üyeler
  SPECIFIC_GRADES = "SPECIFIC_GRADES", // Belirli sınıflar
  SPECIFIC_AGES = "SPECIFIC_AGES", // Belirli yaşlar
  LOYALTY_MEMBERS = "LOYALTY_MEMBERS", // Sadakat üyeleri
  FIRST_TIME_VISITORS = "FIRST_TIME_VISITORS", // İlk kez gelenler
}